refactor(util): convert Position to @import-as-namespace module; rename byPosition to Position.compare
