+++ packages/plugins/plugin-assistant/src/capabilities/companion-chat-provisioner.ts
@@ -14,7 +14,7 @@ import { Obj } from '@dxos/echo';
 import { log } from '@dxos/log';
 import { DeckCapabilities, PLANK_COMPANION_TYPE, type StoredDeckState } from '@dxos/plugin-deck';
 import { getLinkedVariant } from '@dxos/react-ui-attention';
-import { byPosition } from '@dxos/util';
+import { Position } from '@dxos/util';
 
 import { ASSISTANT_COMPANION_VARIANT } from '#meta';
 import { AssistantCapabilities, AssistantOperation } from '#types';
@@ -147,7 +147,7 @@ const resolveEffectiveVariant = (
 ): string | undefined => {
   const companions = Graph.getConnections(graph, plankId, 'child')
     .filter((node) => node.type === PLANK_COMPANION_TYPE)
-    .toSorted((a, b) => byPosition(a.properties, b.properties));
+    .toSorted((a, b) => Position.compare(a.properties, b.properties));
 
   if (companions.length === 0) {
     return undefined;
