+++ packages/plugins/plugin-navtree/src/components/NavTree/NavTree.tsx
@@ -8,7 +8,7 @@ import { useAppGraph } from '@dxos/app-toolkit/ui';
 import { Node, useConnections, useActions as useGraphActions } from '@dxos/plugin-graph';
 import { type MenuItem } from '@dxos/react-ui-menu';
 import { Tabs } from '@dxos/react-ui-tabs';
-import { byPosition } from '@dxos/util';
+import { Position } from '@dxos/util';
 
 import { useLoadDescendents } from '#hooks';
 
@@ -76,11 +76,11 @@ const useTopLevelNavItems = (root?: Node.Node) => {
 
     const topLevelActions = rootActions
       .filter((action) => action.properties.disposition === 'menu')
-      .toSorted((a, b) => byPosition(a.properties, b.properties));
+      .toSorted((a, b) => Position.compare(a.properties, b.properties));
     const pinnedItems = [
       ...outboundPinnedItems,
       ...rootActions.filter((action) => action.properties.disposition === 'pin-end'),
-    ].toSorted((a, b) => byPosition(a.properties, b.properties));
+    ].toSorted((a, b) => Position.compare(a.properties, b.properties));
 
     return {
       pinnedItems,
