+++ packages/sdk/app-toolkit/src/app-graph/AppNode.ts
@@ -271,7 +271,7 @@ export const makeCompanion = <TData = string>({
   label: Translations.Label;
   icon: string;
   data: TData;
-  position?: Position;
+  position?: Position.Position;
 }): Node.NodeArg<TData> => ({
   id,
   type: PLANK_COMPANION_TYPE,
@@ -297,7 +297,7 @@ export const makeDeckCompanion = <TData = any>({
   label: Translations.Label;
   icon: string;
   data: TData;
-  position?: Position;
+  position?: Position.Position;
   joyride?: string;
 }): Node.NodeArg<TData> => ({
   id,
@@ -333,7 +333,7 @@ export const makeSection = ({
   icon: string;
   iconHue?: string;
   space: Space;
-  position?: Position;
+  position?: Position.Position;
   testId?: string;
 }): Node.NodeArg<null> => ({
   id,
@@ -374,7 +374,7 @@ export const makeSettingsPanel = ({
   icon: string;
   /** Hue for the panel's icon. Omit to leave unset (default rendering). */
   iconHue?: string;
-  position?: Position;
+  position?: Position.Position;
 }): Node.NodeArg<string> => ({
   id,
   type,
